# src/cmd/http/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server",
        "title": "API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/tattoos": {
            "post": {
                "description": "```json\n{\n\"tattoos\": [\n{\n\"description\": \"Descripción del tatuaje\",\n\"image\": \"archivo_imagen\",\n\"coord\": {\"x\": 100, \"y\": 200},\n}]\n}\n```",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "tattoos"
                ],
                "summary": "Publicar nuevos tatuajes {admin}",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Imagen del tatuaje (jpg/png/webp)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Formulario inválido o sin datos de tatuajes",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "415": {
                        "description": "El tipo de imagen es inválido, debe ser jpg/png/webp",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "502": {
                        "description": "Falló el repositorio",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/tattoos/latest/{username}": {
            "get": {
                "tags": [
                    "tattoos"
                ],
                "summary": "Recibir los ultimos tatuajes de un usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetTattoosResponse"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/tattoos/{username}": {
            "get": {
                "tags": [
                    "tattoos"
                ],
                "summary": "Recibir los tatuajes de un usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "numero de pagina",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetTattoosResponse"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.GetTattoosResponse": {
            "type": "object",
            "properties": {
                "tattoos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tattoo"
                    }
                }
            }
        },
...
        "model.Coord": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "mime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Profile": {
            "type": "object",
            "properties": {
                "avatar": {
                    "$ref": "#/definitions/model.Image"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idUser": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
...
        "model.Role": {
            "type": "string",
            "enum": [
                "user",
                "tattooArtist",
                "admin"
            ],
            "x-enum-varnames": [
                "USER_ROLE",
                "TATTOO_ARTIST_ROLE",
                "ADMIN_ROLE"
            ]
        },
        "model.Tattoo": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coord": {
                    "$ref": "#/definitions/model.Coord"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idPublication": {
                    "type": "integer"
                },
                "image": {
                    "$ref": "#/definitions/model.Image"
                },
                "likes": {
                    "type": "integer"
                },
                "profile": {
                    "$ref": "#/definitions/model.Profile"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorProblemDetails": {
            "type": "object",
            "properties": {
                "param": {
                    "type": "string",
                    "example": "3"
                },
                "pointer": {
                    "type": "string",
                    "example": "max"
                },
                "title": {
                    "type": "string",
                    "example": "name"
                }
            }
        },
        "utils.ProblemDetails": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Detalle técnico del error"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.ErrorProblemDetails"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Descripción del problema para mostrar al usuario"
                },
                "type": {
                    "type": "string",
                    "example": "/docs/errors/errorPointer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}